Rewards/Scripted Rewards Display Improved The `describeReward()` function used to show scripted (or random) capsule rewards at the top of the mission modal has been rewritten for improved readability, and now breaks up the script rewards into bullet-points, to make them easier to read for the user. It has also been branched off into the `describeRewardIndividual()` function, which works for displaying single-resource rewards.
The `renderMissionButton()` function has been roughly rewritten to improve readability, and for writing to more in the future.
The button next to the mission that users click to access the calculator has now been rewritten so that it can display non-gacha rewards, like Gold and Warps.

Implemented the 'Rare', 'Epic', and 'Supreme' capsule icons to the `missions_img.css` script.

The 'Gold' resource icon has been moved to the 'shared' folder for both games, and it's filepaths have been updated in code.